requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
